[Usability Improvement] Unified Docker setup (Kerberized and non-Kerberized) (#227)
* Add Kerberos docker compose setup

* Actually fix merge conflicts

* Update branch

* Update branch

* Update branch again

* Fix Kerberos example

* Address a few comments

* Fix path for running example in README

* Update README

* Add newline at end of kerberize.sh

* Remove commented out lines in config files

* Change image name to avoid rebuilds

* Extra Docker Automation

* Extra Docker Automation

* fix gitignore

* removed jar

* Ignore keytabs and add newlines at ends of files

* Add instructioins on rebuilding images for Kerberos setup

* Attempt to generate certificate dynamically

* Fix cert generation and address remaining PR comments

* Track keytabs folder

* Update sandbox-clientenv.bat

* Add plugins.sbt for basic-read-example

* Add remaining relevant examples' plugins.sbt files

+++ docker/sandbox-clientenv.bat
@@ -1,6 +1,30 @@
-docker compose up -d
-docker exec docker_hdfs_1 cp /hadoop/conf/core-site.xml /opt/hadoop/etc/hadoop/core-site.xml
-docker exec docker_hdfs_1 cp /hadoop/conf/hdfs-site.xml /opt/hadoop/etc/hadoop/hdfs-site.xml
-docker exec docker_hdfs_1 /opt/hadoop/sbin/stop-dfs.sh
-docker exec docker_hdfs_1 /opt/hadoop/sbin/start-dfs.sh
-docker exec -it docker_client_1 /bin/bash
+@echo off
+
+IF "%~1" == "kerberos" (
+	echo "running kerberos docker compose"
+	docker compose -f docker-compose-kerberos.yml up -d
+	echo "configuring kdc"
+	docker exec kdc /kdc/configure.sh
+	echo "configuring hdfs"
+	docker exec hdfs service ssh start
+	docker exec hdfs start-dfs.sh
+	docker exec -u 0 hdfs /hdfs-krb/kerberize.sh
+	docker exec hdfs stop-dfs.sh
+	docker exec hdfs start-dfs.sh
+	echo "configuring client"
+	docker exec docker_krbclient_1 /client-krb/kerberize.sh
+	docker exec vertica /bin/sh -c "echo %undefined% hdfs.example.com hdfs | sudo tee -a /etc/hosts"
+	docker exec docker_krbclient_1 /bin/sh -c "echo %undefined% hdfs.example.com hdfs | tee -a /etc/hosts"
+	echo "configuring db"
+	docker exec -u 0 vertica /vertica-krb/kerberize.sh
+	docker exec -it docker_krbclient_1 /bin/bash
+) ELSE (
+	echo "running non-kerberized docker compose"
+	docker compose -f docker-compose.yml up -d
+	docker exec docker_hdfs_1 cp /hadoop/conf/core-site.xml /opt/hadoop/etc/hadoop/core-site.xml
+	docker exec docker_hdfs_1 cp /hadoop/conf/hdfs-site.xml /opt/hadoop/etc/hadoop/hdfs-site.xml
+	docker exec docker_hdfs_1 /opt/hadoop/sbin/stop-dfs.sh
+	docker exec docker_hdfs_1 /opt/hadoop/sbin/start-dfs.sh
+	docker exec docker_vertica_1 vsql -c "ALTER DATABASE docker SET MaxClientSessions=100;"
+	docker exec -it docker_client_1 /bin/bash
+)
